requirements.txt
asgiref==3.7.2
Django==5.0.1
djangorestframework==3.14.0
pytz==2024.1
PyYAML==6.0.1
sqlparse==0.4.4
tzdata==2023.4
